+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- rollback : added rollback as alias of revert
+- help : add support for aliases
+
 ### Fixed
 - stop (#024): check the pot existance
 - stop : fixed error messages referring to "jail"